requirements.txt
langchain>=0.2.11
langchain-openai>=0.1.23
langchain-community>=0.2.10
streamlit>=1.36.0
python-dotenv>=1.0.1
pydantic>=2.7.1
requests>=2.32.3
tqdm>=4.66.4
pytest>=8.2.0
ruff>=0.5.0
black>=24.8.0
# Azure optional (for RAG)
azure-search-documents>=11.5.0,<11.6.0
azure-identity>=1.15.0
